Rename copr.py to copr_helper.py to avoid problems
